Fix a bunch of warnings
using the list of warnings from here:
https://github.com/lefticus/cppbestpractices/blob/e73393f25a85f83fed7399d8b65cb117d00b2231/02-Use_the_Tools_Available.md#L100

Seems like a good place to start, and would improve things a bit
type-wise.  This patchset attempts to correct all the issues in one
shot.

Tested:
It builds.  Will test various subsystems that have been touched

+++ include/persistent_data_middleware.hpp
@@ -25,7 +25,7 @@ namespace fs = std::filesystem;
 
 class Middleware
 {
-    int jsonRevision = 1;
+    uint64_t jsonRevision = 1;
 
   public:
     // todo(ed) should read this from a fixed location somewhere, not CWD
@@ -62,7 +62,7 @@ class Middleware
     void readData()
     {
         std::ifstream persistentFile(filename);
-        int fileRevision = 0;
+        uint64_t fileRevision = 0;
         if (persistentFile.is_open())
         {
             // call with exceptions disabled
